C program to implement a LC‐3 simulator
Implements the instructions LD, LEA, LDI, AND, NOT and ADD from a
hexadecimal .obj file.

For this part, the simulator displays the hexadecimal code of each BR
instruction that has been executed and the state of the LC‐3 machine
after each BR instruction is executed and implements the other
instructions.